Move PurchaseHistory rows to OrderRow component
